feat: add Docker build and release call workflow Action`
Add a workflow that takes a Dockerfile, target, a registry, and desired
platforms, builds images, and pushes them to the registry, and creates
attestations.

Keeps some of the inputs from the Goreleaser workflow, but not all. I
removed some stuff that looked unecessary (because it could be done in
the Dockerfile instead) or where it looked like there were better
options for reasonable defaults.

If the pushed ref is a tag, create a release for it, unless the release
already exists.

Signed-off-by: Travis Raines <571832+rainest@users.noreply.github.com>